ref(metrics): move metrics api query endpoint to separate file
